// src/plugins/nobrinde/customer-support/gql/generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: unknown; output: unknown; }
  JSON: { input: unknown; output: unknown; }
  Money: { input: number; output: number; }
  Upload: { input: unknown; output: unknown; }
};
// ...
export type AddSupportTicketMessageInput = {
  content: Scalars['String']['input'];
  supportTicketId: Scalars['ID']['input'];
};
// ...
export type CreateSupportTicketInput = {
  channelId: Scalars['String']['input'];
  description: Scalars['String']['input'];
  priority?: InputMaybe<SupportTicketPriority>;
  subjectId: Scalars['String']['input'];
};
// ...
export enum SupportTicketPriority {
  HIGH = 'HIGH',
  LOW = 'LOW',
  MEDIUM = 'MEDIUM',
  URGENT = 'URGENT'
}

export enum SupportTicketSender {
  CUSTOMER = 'CUSTOMER',
  SELLER = 'SELLER'
}

export type SupportTicketSortParameter = {
  channelId?: InputMaybe<SortOrder>;
  createdAt?: InputMaybe<SortOrder>;
  customerId?: InputMaybe<SortOrder>;
  description?: InputMaybe<SortOrder>;
  id?: InputMaybe<SortOrder>;
  subjectId?: InputMaybe<SortOrder>;
  updatedAt?: InputMaybe<SortOrder>;
};

export enum SupportTicketStatus {
  CLOSED = 'CLOSED',
  OPEN = 'OPEN',
  PENDING = 'PENDING'
}
// ...
export type UpdateSupportTicketInput = {
  channelId?: InputMaybe<Scalars['String']['input']>;
  description?: InputMaybe<Scalars['String']['input']>;
  id: Scalars['ID']['input'];
  priority?: InputMaybe<SupportTicketPriority>;
  status?: InputMaybe<SupportTicketStatus>;
  subjectId?: InputMaybe<Scalars['String']['input']>;
};

export type UpdateSupportTicketStatusInput = {
  status: SupportTicketStatus;
  supportTicketId: Scalars['ID']['input'];
};